Actualizaciones de rutas de endpoint, se crea controller para recuperar y resetear contraseña
