+++ src/config/dbConnect.js
@@ -23,6 +23,14 @@ async function dbConnect() {
     cached.promise = mongoose.connect(MONGODB_URI).then((mongoose) => {
       console.log('Conectado a la base de datos');
       return mongoose;
+    }).catch((error) => {
+      if (error.code === 'ETIMEOUT') {
+        console.error('Error de conexión: Tiempo de espera agotado.');
+        console.error('Por favor, verifique que su IP pública esté configurada en MongoDB Atlas para permitir la conexión a la base de datos.');
+      } else {
+        console.error('Error al conectar con la base de datos:', error.message);
+      }
+      throw error;
     });
   }
