+++ src/controllers/loginController.js
@@ -1,6 +1,6 @@
 import jwt from 'jsonwebtoken';
 import dbConnect from '../config/dbConnect.js';
-import User from '../models/User.js';
+import User from '../models/UserModel.js';
 import bcrypt from 'bcryptjs';
 
 // Clave secreta para firmar el token (asegúrate de definir esta variable en tus variables de entorno)
